requirements.txt
pandas
joblib
mysql-connector-python
scikit-learn
flask